Refactor cordova native code
